% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle1S-package.R
\docType{package}
\name{eagle1S-package}
\alias{eagle1S-package}
\alias{eagle1S}
\title{The 'eagle1S' package.}
\description{
Re-implementation of EAGLE using Stan. Detect GxE QTLs using allele specific expression.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
