% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle1S.R
\name{eagle1S}
\alias{eagle1S}
\title{Map eQTLs and gxeQTLs using allele specific expression}
\usage{
eagle1S(ase, phased, meta, checkpoint_dir = NULL, cisdist = 1e+05,
  min_samples = 10, concShape = 1.001, concRate = 0.001,
  min_reads = 1000, max_iter = 30)
}
\arguments{
\item{ase}{A table with SNP position ("POS"), alternative allele ("ALT"), "individual", reference counts "r" and alternative allele counts "a"}

\item{phased}{A VCF table, i.e. genotypes should start at column 11. Should have phased genotypes e.g. 0|1. Missing values, and equivalent unphased values, are allowed and will be ignored. Column names must correspond to the individuals and be compatible with `ase` and `meta` tables.}

\item{meta}{A table with an "individual" column and environmental factor "x". x should be normalized to some sensible scaling.}
}
\value{
Tibble with log likelihoods for all tested SNP pairs.
}
\description{
Map eQTLs and gxeQTLs using allele specific expression
}
