% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle1S.R
\name{get_lead_eQTL}
\alias{get_lead_eQTL}
\title{Get lead eQTLs}
\usage{
get_lead_eQTL(results_table)
}
\arguments{
\item{results_table}{from `eagle1S`}
}
\value{
Tibble with per exonic SNP top/lead eSNPs
}
\description{
Get lead eQTLs
}
