% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle1S.R
\name{get_lead_gxeQTL}
\alias{get_lead_gxeQTL}
\title{Get lead gxeQTLs}
\usage{
get_lead_gxeQTL(results_table)
}
\arguments{
\item{results_table}{from `eagle1S`}
}
\value{
Tibble with per exonic SNP top/lead gxeSNPs
}
\description{
Get lead gxeQTLs
}
